refactor(edge): extract control plane bootstrap wiring from listener
