+++ crates/edge/src/quic_listener/startup.rs
@@ -369,79 +369,6 @@ impl QUICListener {
         })
     }
 
-    pub fn spawn_control_plane_tasks(
-        config: &RuntimeConfig,
-        shared_state: &SharedRuntimeState,
-        worker_count: usize,
-    ) -> Result<(), ProxyError> {
-        shared_state
-            .watchdog
-            .set_expected_workers(worker_count.max(1));
-        Self::spawn_generation_background_tasks(config, shared_state);
-        Self::spawn_metrics_endpoint(config, Arc::clone(&shared_state.metrics), None)?;
-        Self::spawn_control_api_endpoint(config, shared_state, None, worker_count)?;
-        Ok(())
-    }
-
-    pub fn spawn_control_plane_tasks_with_runtime_bundle(
-        config: &RuntimeConfig,
-        shared_state: &SharedRuntimeState,
-        runtime_bundle: Arc<RuntimeBundleHandle>,
-        worker_count: usize,
-    ) -> Result<(), ProxyError> {
-        shared_state
-            .watchdog
-            .set_expected_workers(worker_count.max(1));
-        Self::spawn_generation_background_tasks(config, shared_state);
-        Self::spawn_metrics_endpoint(
-            config,
-            Arc::clone(&shared_state.metrics),
-            Some(Arc::clone(&runtime_bundle)),
-        )?;
-        Self::spawn_control_api_endpoint(config, shared_state, Some(runtime_bundle), worker_count)?;
-        Ok(())
-    }
-
-    pub(super) fn spawn_generation_background_tasks(
-        config: &RuntimeConfig,
-        shared_state: &SharedRuntimeState,
-    ) {
-        shared_state.watchdog.set_expected_workers(
-            config
-                .policies
-                .transport
-                .worker_threads
-                .max(1)
-                .saturating_mul(config.policies.transport.packet_shards_per_worker.max(1))
-                .max(1),
-        );
-        let task_registry = Arc::clone(&shared_state.generation_tasks);
-        Self::spawn_backend_dns_refresh(
-            config,
-            Arc::clone(&shared_state.transport_pool),
-            Arc::clone(&shared_state.backend_resolution_store),
-            shared_state.backend_dns_resolver.clone(),
-            Arc::clone(&shared_state.metrics),
-            Arc::clone(&task_registry),
-        );
-        Self::spawn_health_checks(
-            shared_state.upstream_pools.clone(),
-            Arc::clone(&shared_state.transport_pool),
-            Arc::clone(&shared_state.backend_endpoints),
-            Arc::clone(&shared_state.backend_health_checks),
-            Arc::clone(&shared_state.backend_resolution_store),
-            Arc::clone(&shared_state.metrics),
-            Arc::clone(&task_registry),
-        );
-        Self::spawn_watchdog(
-            config,
-            Arc::clone(&shared_state.metrics),
-            Arc::clone(&shared_state.resilience),
-            Arc::clone(&shared_state.watchdog),
-            task_registry,
-        );
-    }
-
     pub fn bind_reuseport_sockets(
         config: &ListenerRuntimeConfig,
         workers: usize,
